# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /groceries/add:
    post:
      operationId: GroceriesController_addRemoveMealToGroceries
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMealToGroceriesRequestDto'
      responses:
        '201':
          description: ''
      tags: &ref_2
        - Groceries
...
  /groceries:
    get:
      operationId: GroceriesController_getUserGroceries
      parameters:
        - name: date
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserGroceriesResponseDto_Output'
      tags: *ref_2
    post:
      operationId: GroceriesController_toggleIngredientsStatus
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToggleIngredientsStatusRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToggleIngredientsStatusResponseDto_Output'
      tags: *ref_2
...
  /meal-plans:
    get:
      operationId: MealPlanController_getMealPlan
      parameters:
        - name: date
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: &ref_4
        - MealPlan
    post:
      operationId: MealPlanController_createMealPlanEntries
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMealPlanRequestDto'
      responses:
        '201':
          description: ''
      tags: *ref_4
  /meal-plans/{id}:
    delete:
      operationId: MealPlanController_deleteMealPlanEntry
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
  /meal-plans/{id}/servings:
    patch:
      operationId: MealPlanController_updateServings
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMealPlanServingsRequestDto'
      responses:
        '200':
          description: ''
      tags: *ref_4
info:
  title: MonAmiChef
  description: The MonAmiChef API documentation
  version: '1.0'
  contact: {}
...
servers:
  - url: https://api.monamichef.com
    description: Production server
  - url: http://localhost:8888
    description: Local dev
components:
  securitySchemes:
    X-API-KEY:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    AddMealToGroceriesRequestDto:
      type: object
      properties:
        recipeId:
          type: string
        newState:
          type: boolean
      required:
        - recipeId
        - newState
    GetUserGroceriesResponseDto_Output:
      type: object
      properties:
        root:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              totalQuantity:
                type: number
              unit:
                type: string
              category:
                type: string
              recipeDetails:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    servings:
                      type: number
                  required:
                    - id
                    - name
                    - servings
                  additionalProperties: false
              isBought:
                type: boolean
              ingredientIds:
                type: array
                items:
                  type: string
              image:
                anyOf:
                  - type: string
                  - type: 'null'
            required:
              - name
              - totalQuantity
              - unit
              - category
              - recipeDetails
              - isBought
              - ingredientIds
              - image
            additionalProperties: false
          x-nestjs_zod-unwrap-root: true
          x-nestjs_zod-has-null: true
    ToggleIngredientsStatusRequestDto:
      type: object
      properties:
        ingredientIds:
          type: array
          items:
            type: string
        isBought:
          type: boolean
      required:
        - ingredientIds
        - isBought
    ToggleIngredientsStatusResponseDto_Output:
      type: object
      properties:
        count:
          type: number
          x-nestjs_zod-parent-additional-properties: false
      required:
        - count
...
    CreateMealPlanRequestDto:
      type: object
      properties:
        entries:
          minItems: 1
          type: array
          items:
            type: object
            properties:
              recipeId:
                type: string
              date:
                type: string
              mealType:
                type: string
                enum:
                  - BREAKFAST
                  - LUNCH
                  - DINNER
              servings:
                default: 1
                type: integer
                minimum: 1
                maximum: 9007199254740991
            required:
              - recipeId
              - date
              - mealType
      required:
        - entries
    UpdateMealPlanServingsRequestDto:
      type: object
      properties:
        servings:
          type: integer
          minimum: 1
          maximum: 9007199254740991
      required:
        - servings
security:
  - X-API-KEY: []